Help Window overhaul
Now it looks way nicer, actually tells you all the hotkeys in the game, and the Tiles tab has been replaced with FAQ.
Hopefully I'm not missing anything. If I am, do tell.
(Also I rearranged the scene in a way that tileset_scroll.gd isn't needed anymore. Two other scenes use it, so it can't be deleted, but you should be able to change them in a way for it to be unneeded.)