add timer manager
this allows u to create any number of arbitrary timers to be shown on the hud

first, get the timer manager by calling get_timer_manager() on the player scene

then, call add_timer() on the timer manager. it'll return the timer it creates!

you can then connect the "time_over" signal on the returned timer to do whatever u please with :)